+++ sysarq/src/pages/FieldsRegister/BoxAbbreviation.js
@@ -3,5 +3,5 @@ import React from "react";
 import DataTable from "../components/DataTable";
 
 export default function BoxAbbreviation() {
-	return <DataTable title="Sigla da Caixa" url="box_abbreviation/" />;
+	return <DataTable title="Sigla da Caixa" url="box-abbreviation/" />;
 }
