+++ sysarq/src/pages/FieldsRegister/CreateBoxAbbreviation.js
@@ -1,8 +1,7 @@
-import React, { useState } from "react";
-import Alert from "@material-ui/lab/Alert";
-import { Paper, TextField, Grid, Container } from "@material-ui/core";
+import { useState } from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import { axiosArchives } from "../../Api";
+import { createForm2 } from "./form";
 
 const useStyles = makeStyles((theme) => ({
 	input: {
@@ -131,49 +130,14 @@ export default function CreateBoxAbbreviation() {
 	const title = "Arquivo Geral da Policia Civil de Goiás";
 	const subtitle = "Cadastrar caixa";
 
-	return (
-		<div className="create-form-container">
-			{show === true ? <Alert severity="success">Campo cadastrado!</Alert> : ""}
-			{showError === true ? (
-				<Alert severity="error">Erro de conexão!</Alert>
-			) : (
-				""
-			)}
-			<Paper className="form-cadastro-container" elevation={10}>
-				<h1>{title}</h1>
-				<h2>{subtitle}</h2>
-				<div className="inputs-container">
-					<Container className="container">
-						<Grid container spacing={2}>
-							{fields.map((item, key) => {
-								const input = (
-									<Grid item xs={12} sm={12} md={12} key={key.toString()}>
-										<TextField
-											id={item.placeholder}
-											label={item.placeholder}
-											type={item.type}
-											value={item.value}
-											onChange={(event) => {
-												item.setValue(event.target.value);
-												item.setHelperText("");
-												item.setError(false);
-											}}
-											className={classes.input}
-											helperText={item.helperText}
-											error={item.error}
-											inputProps={{ maxLength: "100" }}
-										/>
-									</Grid>
-								);
-								return input;
-							})}
-						</Grid>
-					</Container>
-				</div>
-				<button data-testid="click" type="button" onClick={onClick}>
-					CADASTRAR
-				</button>
-			</Paper>
-		</div>
+	return createForm2(
+		fields,
+		title,
+		subtitle,
+		classes,
+		show,
+		showError,
+		onClick,
+		true
 	);
 }
