+++ sysarq/src/pages/FieldsRegister/DocumentSubject.js
@@ -3,5 +3,5 @@ import React from "react";
 import DataTable from "../components/DataTable";
 
 export default function DocumentSubject() {
-	return <DataTable title="Assunto do Documento" url="document_subject/" />;
+	return <DataTable title="Assunto do Documento" url="document-subject/" />;
 }
