+++ sysarq/src/pages/FieldsRegister/DocumentType.js
@@ -3,5 +3,5 @@ import React from "react";
 import DataTable from "../components/DataTable";
 
 export default function DocumentType() {
-	return <DataTable title="Tipo de Documento" url="document_type/" />;
+	return <DataTable title="Tipo de Documento" url="document-type/" />;
 }
